Added a build script to prepare for a deployment.

+++ dev-bin/build.sh
@@ -1,5 +1,33 @@
+#!/bin/sh
+
+SPECIFIED_VERSION="$1"
+
 clear
 
+if [ "`git status -s`" ]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
+
+if [ -z "$SPECIFIED_VERSION" ]
+then
+  echo "You must specify a version number"
+  exit 1;
+fi
+
+echo "Building OpenMud Version: $SPECIFIED_VERSION"
+
+# Get the current GIT branch
+CURRENT_BRANCH_NAME=$(git symbolic-ref -q HEAD)
+CURRENT_BRANCH_NAME=${CURRENT_BRANCH_NAME##refs/heads/}
+CURRENT_BRANCH_NAME=${CURRENT_BRANCH_NAME:-HEAD}
+
+echo "Checking out version tag"
+
+git checkout $SPECIFIED_VERSION --quiet
+
+# Get the version number specified in the package.json file
 PACKAGE_VERSION=$(cat package.json \
   | grep version \
   | head -1 \
@@ -7,7 +35,11 @@ PACKAGE_VERSION=$(cat package.json \
   | sed 's/[",]//g' \
   | tr -d '[[:space:]]')
 
-echo "Building OpenMud $PACKAGE_VERSION"
+if [[ "$PACKAGE_VERSION" != "$SPECIFIED_VERSION" ]]
+then
+  echo "Your package.json file version number does not match the git tag."
+  exit 1;
+fi
 
 echo "Cleaning up previous builds"
 rm -rf dist/
@@ -26,3 +58,7 @@ cd ..
 
 echo "Cleaning temporary files"
 rm -r dist/
+
+echo "Returning to original GIT branch"
+
+git checkout $CURRENT_BRANCH_NAME --quiet
